// orbito-frontend/src/core/api/generated/models/teamMemberRole.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Orbito API
 * OpenAPI spec version: v1
 */

export type TeamMemberRole = typeof TeamMemberRole[keyof typeof TeamMemberRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TeamMemberRole = {
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
} as const;